fix 64bit warning

+++ external/include/osc/udp.h
@@ -282,9 +282,9 @@ struct UdpSocket {
     do {
       int res;
       if (isBound()) {
-        res = sendto(handle, (const char*)ptr, sz, 0, &addr.addr(), (int)addr.actualLen());
+        res = (int)sendto(handle, (const char*)ptr, sz, 0, &addr.addr(), (int)addr.actualLen());
       } else {
-        res = send(handle, (const char*)ptr, (int)sz, 0);
+        res = (int)send(handle, (const char*)ptr, (int)sz, 0);
         //        res = write(handle, ptr, sz);
       }
 #ifdef WIN32
